Cleanup cooldown code for actionv2
