+++ UnityProject/Assets/Scripts/Actions/V2/ActionManager.cs
@@ -20,32 +20,6 @@ namespace Actions.V2
 
 		private const float MINIMUM_COOLDOWN_TIME = 0.085f;
 
-		[Serializable]
-		public class CooldownInfo : NetworkMessage
-		{
-			public string ActionId { get; private set; }
-			public DateTime CooldownEnd { get; private set; }
-
-			public CooldownInfo() { }
-			public CooldownInfo(string actionId, DateTime cooldownEnd)
-			{
-				ActionId = actionId;
-				CooldownEnd = cooldownEnd;
-			}
-
-			public void Serialize(NetworkWriter writer)
-			{
-				writer.WriteString(ActionId);
-				writer.WriteLong(CooldownEnd.Ticks);
-			}
-
-			public void Deserialize(NetworkReader reader)
-			{
-				ActionId = reader.ReadString();
-				CooldownEnd = new DateTime(reader.ReadLong());
-			}
-		}
-
 		public enum OnType
 		{
 			Body,
