+++ UnityProject/Assets/Scripts/Actions/V2/Cooldown.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using Mirror;
+
+namespace Actions.V2
+{
+	[Serializable]
+	public class CooldownInfo : NetworkMessage
+	{
+		public string ActionId { get; private set; }
+		public DateTime CooldownEnd { get; private set; }
+
+		public CooldownInfo() { }
+		public CooldownInfo(string actionId, DateTime cooldownEnd)
+		{
+			ActionId = actionId;
+			CooldownEnd = cooldownEnd;
+		}
+
+		public void Serialize(NetworkWriter writer)
+		{
+			writer.WriteString(ActionId);
+			writer.WriteLong(CooldownEnd.Ticks);
+		}
+
+		public void Deserialize(NetworkReader reader)
+		{
+			ActionId = reader.ReadString();
+			CooldownEnd = new DateTime(reader.ReadLong());
+		}
+	}
+}
