+++ UnityProject/Assets/Scripts/Actions/V2/Cooldown.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 032a0441a14d4a91bc62cf9d252b3097
+timeCreated: 1766740793
